Less regenerating playlist filter and shuffle lists in particular at startup. Faster shuffle generation. hotkeys for apps dont require atoms. Lots of ints changed to unsigned longs so all features support the same max file count.

+++ Tuniac1/TuniacApp/IPlaylist.h
@@ -86,8 +86,8 @@ public:
 	virtual bool	GetTextRepresentation(unsigned long ulFieldID, LPTSTR szString, unsigned long ulNumChars)		= 0;
 };
 
-typedef Array<int, 100>						IndexArray;
-typedef Array<IPlaylistEntry *, 100>		EntryArray;
+typedef Array<unsigned long, 100>		IndexArray;
+typedef Array<IPlaylistEntry *, 100>	EntryArray;
 
 class IPlaylist
 {
